feat: update GetOperatorDefinition functions to support dynamic definition (#47)
Because

Originally, our connector definitions are fixed. However, in some
scenarios, the definition will be dynamic.

The component JSON schema can be influenced by connector configuration.
For example, the Instill Model Connector can retrieve the model list and
incorporate it into the component JSON schema as an enumeration.

The OpenAPI schema can be influenced by both connector and component
configuration. For example, we can let user define the JSON schema of
the output body in the RestAPI connector and utilize it to generate the
OpenAPI schema of the RestAPI connector.

Therefore, we would like to make these connector definitions dynamically
generated.

This commit

- update GetOperatorDefinition functions to support dynamic definition