feat: introduce `instillFormat: semi-structured/json` (#55)
Because

- In our current implementation, the `instillFormat:
semi-structured/object` only allows for `object` data. However, in some
scenarios, we would like to define a field as JSON. JSON data can be an
object, array, number, string, or boolean.

This commit

- Introduce `instillFormat: semi-structured/json`; this instillFormat
will be treated as a free-form data type, which means we won't have a
`type` in the JSON schema.