feat: simplify `openapi_specifications` to `data_specifications` (#64)
Because

- Related to this https://github.com/instill-ai/protobufs/pull/282,
currently, we have an `openapi_specification` field in the
connector/operator definition. The original purpose was that we could
generate an OpenAPI schema and have some UI to let the user debug on an
individual component. However, it turns out that we don't really need
that OpenAPI schema, and it makes our response very large and hard to
read.

This commit

- Simplifies `openapi_specification` into `data_specification` by
preserving only the input and output JSON schema and removing the nested
structure of the OpenAPI spec.