feat: merge connector and operator repos into this repo (#72)
Because
- As our project expands to include various component types such as
generic, iterator, and AI components, managing separate repositories for
each becomes less practical. Although the initial plan aimed to
integrate them directly into the component repository, the introduction
of the new AI Component poses challenges. Given that the AI Component
relies on the implementation of all connectors in the connector
repository, including it in the component repository would lead to
cyclic dependencies. Therefore, the optimal solution is to consolidate
all components into a single repository.

This commit

- Merges code from [connector](https://github.com/instill-ai/connector)
and [operator](https://github.com/instill-ai/operator) into this
repository.
- Removes unused `objectmapper`.

---------

Co-authored-by: Juan Vallés <3977183+jvallesm@users.noreply.github.com>