feat: refactor package structure (#140)
Because

- In our new VDP recipe design, there is not much difference between
connectors and operators, and the term "connector" doesn't provide
useful information to the user. Therefore, we decided to flatten the
component structure. The original connectors will be directly
categorized into `ai`, `data`, and `application` components.

This commit

- Modifies the folder structure to categorize connectors into `ai`,
`data`, and `application` components.
- Removes `/pkg` folder.

Note

- In this PR, we only modify the folder structure. In the upcoming PRs,
we'll further refactor the component interface to unify the connector
and operator interfaces into one component interface.