feat: unify component interface (#144)
Because

- We decided to remove the connector and operator categorization and
organize the components into AI, data, application, and operator
components. The original `IConnector` and `IOperator` interfaces are no
longer needed.
- We had a `connection` field for the connectors to set up the
connection. To make this term more general, we renamed it to `setup`.
The `setup` field should be applicable to all kinds of components. For
example, for a CPU-bound operator, we might need to configure some
hardware resources in the future. This configuration can also be placed
in the `setup` field. Additionally, to improve the readability of the
JSON file, we moved the configuration into a new `setup.json` file.
- We are going to recategorize `googlesearch`, `redis`, `restapi`, and
`website` into the "application" component.

This commit

- Removes the base connector and operator interfaces and structs. Moves
all functions to the base component struct.
- Moves all `connection` properties into a new `setup.json` file.
- Bumps the component version of `googlesearch`, `redis`, `restapi`, and
`website` to make the component type updated.
- Moves the `component.go` file into a new package `/store`.