feat(jira): add action tasks (#241)
Because

- We need actions to complete the user story

This commit

- Provide a draft json schema of each task

8/12 Update
I have implemented all functions to connect to the real API and found
some tasks are related to some APIs that are not compatible with the
current version of Jira. The remaining tasks are:
- TASK_CREATE_ISSUE
- TASK_UPDATE_ISSUE
- TASK_CREATE_SPRINT
- TASK_UPDATE_SPRINT

--------------
7/31 Update
@chuang8511 Hi, I added three more tasks and moved the schema from v1 to
v0 as #205 is merged. There are 5 new tasks now:
- TASK_CREATE_ISSUE
- TASK_UPDATE_ISSUE
- TASK_CREATE_SPRINT
- TASK_UPDATE_SPRINT
- TASK_RANK_EPIC

--------------
7/25
Note that only the last 3 tasks need to be reviewed, as the others are
under review at #205
- TASK_CREATE_ISSUE
- TASK_UPDATE_ISSUE
- TASK_MOVE_ISSUE_TO_EPIC