fix(compogen): fix wrong anchors and missing object descriptions (#356)
Because

- the auto-gen `oneOf` objects mess up the TOC of the documentation
website due to wrong anchor syntax
- the object descriptions are missing
- `make build-doc` should use local codebase to build and install or
otherwise the local change of `tools/compogen` can't be applied to
auto-gen the latest component `README.mdx`.

This commit

- fix the issues mentioned above
- remove redundant empty lines in the readme template
- implement the precedence logic of task and `oneOf` object
`description` and `instillShortDescription`
- `description` and `instillShortDescription` are used by the console to
provide a description of the task/object in the component. If
`instillShortDescription` does not exist, it will be the same as
`description`.